\name{loadPed}
\alias{loadPed}
\title{Loads in genotype data from a PED and MAP file.}
\usage{
loadPed(fileprefix)
}
\arguments{
  \item{fileprefix}{prefix of the PED and MAP file to
  load.}
}
\value{
A matrix of genotypes in 0,1,2 coding
}
\description{
Loads in genotype data from a PED and MAP file.
}

