% Generated by roxygen2 (4.0.0): do not edit by hand
\name{maf_from_dosage}
\alias{maf_from_dosage}
\title{Calculate Minor Allele Frequency from dosage levels}
\usage{
maf_from_dosage(x)
}
\arguments{
\item{x}{vector of dosages as given by `loadPed`.}
}
\value{
the frequency of a minor allele in your vector.
}
\description{
Useful for calculating the MAF for subsets of individuals
}

